+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+[ToggleSheet] - Fix bottom insets and safe area bottom padding rendering issues
+[MultiStateSheet] - Make footer `HitBehavior` default to `null` if its placed on top of content list
+
 ## 1.0.0+1
 [Package] - format files to the dart standart